day 2 repackage
