+++ src/main/java/com/controller/AddProductServlet.java
@@ -1,8 +1,8 @@
-package controller;
+package com.controller;
 
-import factory.ProductServiceFactory;
-import model.Product;
-import service.ProductService;
+import com.factory.ProductServiceFactory;
+import com.model.Product;
+import com.service.ProductService;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,15 +11,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(value = "/add")
-public class ProductAddServlet extends HttpServlet {
+@WebServlet(value = "/addProduct")
+public class AddProductServlet extends HttpServlet {
 
     private static final ProductService productService = ProductServiceFactory.getInstance();
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        req.getRequestDispatcher("productAdd.jsp").forward(req, resp);
+        req.getRequestDispatcher("addProduct.jsp").forward(req, resp);
     }
 
     @Override
@@ -29,9 +29,7 @@ public class ProductAddServlet extends HttpServlet {
         String description = req.getParameter("description");
         Double price = Double.parseDouble(req.getParameter("price"));
 
-        Product product = productService.create(name, description, price);
-
-        productService.add(product);
+        productService.addProduct(name, description, price);
         resp.setStatus(HttpServletResponse.SC_OK);
         req.getRequestDispatcher("users.jsp").forward(req, resp);
     }
