+++ src/main/java/com/main/ShopApplication.java
@@ -1,4 +1,4 @@
-package main;
+package com.main;
 
 public class ShopApplication {
