+++ src/main/java/com/model/Product.java
@@ -1,4 +1,4 @@
-package model;
+package com.model;
 
 public class Product {
