+++ src/main/java/com/model/User.java
@@ -1,4 +1,4 @@
-package model;
+package com.model;
 
 public class User {
