--- src/main/java/service/ProductService.java
@@ -1,14 +1,0 @@
-package service;
-
-import model.Product;
-
-import java.util.List;
-
-public interface ProductService {
-
-    Product create(String name, String description, Double price);
-
-    void add(Product item);
-
-    List<Product> getAll();
-}
